+++ libs/attachment-canvas/src/types/attachment-canvas.ts
@@ -1,0 +1,6 @@
+/** The type of content the canvas can display. */
+export enum AttachmentContentType {
+  PlainText = 'plain_text',
+  Image = 'image',
+  Unsupported = 'unsupported',
+}
